requirements.txt
python-dotenv~=1.2.1
pyserial~=3.5
mysql-connector-python~=9.5.0